# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SWARM Rendezvous API",
    "description": "Rendezvous service for the SWARM P2P media suite: accounts, swarms, join codes, device registry, and WSS signaling (see /api/v1/ws — not representable in OpenAPI).",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices/{device_id}/metadata": {
      "patch": {
        "tags": [
          "devices"
        ],
        "operationId": "patch_metadata",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/swarms": {
      "get": {
        "tags": [
          "swarms"
        ],
        "operationId": "list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwarmListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "swarms"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSwarmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwarmSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/swarms/join": {
      "post": {
        "tags": [
          "devices"
        ],
        "operationId": "join_swarm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinSwarmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwarmSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/swarms/{swarm_id}/codes": {
      "post": {
        "tags": [
          "swarms"
        ],
        "operationId": "create_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinCodeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/swarms/{swarm_id}/devices/{device_id}": {
      "delete": {
        "tags": [
          "swarms"
        ],
        "summary": "A device leaves one swarm it belongs to, without touching its other\nmemberships or its access token — symmetric with `devices::join_swarm`.\nSelf-only (mirrors `devices::patch_metadata`'s own-resource check): a\ndevice cannot be kicked from a swarm by another device, only by the\nswarm's owner deleting the whole swarm (`delete`, above).",
        "operationId": "leave",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateCodeRequest": {
        "type": "object",
        "properties": {
          "device_type_hint": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DeviceType",
                "description": "Optional expected device type, shown in the UI; not enforced at\nredemption (the device declares its real type)."
              }
            ],
            "default": null
          }
        },
        "additionalProperties": false
      },
      "CreateSwarmRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DeviceType": {
        "type": "string",
        "enum": [
          "client",
          "server",
          "both"
        ]
      },
      "JoinCodeResponse": {
        "type": "object",
        "required": [
          "code",
          "expires_at"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Shown exactly once; only a hash is stored."
          },
          "expires_at": {
            "type": "string"
          }
        }
      },
      "JoinSwarmRequest": {
        "type": "object",
        "description": "`POST /api/v1/swarms/join` — add an already-registered device (Bearer auth)\nto another swarm with a fresh join code.",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "MetadataPatchRequest": {
        "type": "object",
        "description": "`PATCH /api/v1/devices/{device_id}/metadata` — a device updating its own\narbitrary key/value metadata (e.g. a server self-reporting the address\npeers should dial: `{\"peer_addr\": \"192.168.1.50:8543\"}`).",
        "required": [
          "metadata"
        ],
        "properties": {
          "metadata": {
            "type": "object",
            "description": "Keys set to an empty string are removed; others upserted.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "SwarmListEntry": {
        "type": "object",
        "required": [
          "id",
          "name",
          "device_count"
        ],
        "properties": {
          "device_count": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "SwarmListResponse": {
        "type": "object",
        "required": [
          "swarms"
        ],
        "properties": {
          "swarms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SwarmListEntry"
            }
          }
        }
      },
      "SwarmSummary": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "swarm_session"
      }
    }
  },
...
}